+++ crates/zizmor/src/audit/hardcoded_container_credentials.rs
@@ -27,7 +27,7 @@ impl Audit for HardcodedContainerCredentials {
 
     fn audit_workflow<'doc>(
         &self,
-        workflow: &'doc crate::models::Workflow,
+        workflow: &'doc crate::models::workflow::Workflow,
     ) -> anyhow::Result<Vec<crate::finding::Finding<'doc>>> {
         let mut findings = vec![];