+++ crates/zizmor/src/models/coordinate.rs
@@ -324,11 +324,12 @@ pub(crate) enum Usage {
 mod tests {
     use std::str::FromStr;
 
+    use github_actions_expressions::context;
     use github_actions_models::workflow::job::Step;
 
     use super::{ActionCoordinate, StepCommon};
     use crate::{
-        finding::location::Locatable,
+        finding::location::{Locatable, SymbolicLocation},
         models::{
             coordinate::{ControlExpr, ControlFieldType, Toggle, Usage},
             uses::RepositoryUsesPattern,
@@ -337,7 +338,7 @@ mod tests {
 
     // Test-only trait impls.
     impl<'doc> Locatable<'doc> for Step {
-        fn location(&self) -> crate::models::SymbolicLocation<'doc> {
+        fn location(&self) -> SymbolicLocation<'doc> {
             unreachable!()
         }
 
@@ -351,7 +352,7 @@ mod tests {
             unreachable!()
         }
 
-        fn env_is_static(&self, _name: &str) -> bool {
+        fn env_is_static(&self, _ctx: &context::Context) -> bool {
             unreachable!()
         }