+++ docs/release-notes.md
@@ -17,6 +17,8 @@ of `zizmor`.
   action definitions, rather than just workflow definitions (#899)
 * The [bot-conditions] audit now detects more spoofable actor checks,
   including checks against well-known user IDs for bot accounts (#905)
+* The [template-injection] and other audits now produce more precise
+  results when analyzing `env` context accesses for static-ness (#911)
 
 ### Bug Fixes 🐛
 
@@ -25,6 +27,8 @@ of `zizmor`.
   composite action `uses:` step (#887)
 * The [bot-conditions] audit now correctly analyzes index-style contexts,
   e.g. `github['actor']` (#905)
+* Fixed a bug where `zizmor` would fail to parse expressions that
+  contained `>=` or `<=` (#916)
 
 ## 1.9.0